refactor: move conversation title & image inside virtual scroll area
